requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
librt==0.9.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy==1.20.2
mypy_extensions==1.1.0
numpy==2.4.4
oauthlib==3.3.1
pandas==3.0.2
pathspec==1.1.0
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
